user, kernel: use 32 bit IPCP addresses

+++ include/rlite/common.h
@@ -32,7 +32,7 @@ struct rina_name {
     char *aei;
 } __attribute__((packed));
 
-typedef uint64_t rl_addr_t;
+typedef uint32_t rl_addr_t;
 typedef uint32_t rl_port_t;
 typedef uint16_t rl_ipcp_id_t;
 typedef uint16_t rl_msg_t;
